Refactoring: API 함수 정리
